requirements.txt
Django==4.2
djangorestframework==3.14
drf-spectacular==0.28
psycopg2-binary
tensorflow==2.14.0
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.0
torch==2.0.1
torchvision==0.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
lightgbm

